Initial commit (although I should have been doing this for a while now).
So far this app will allow the user to manage a wait list for a restaurant.
This wait list is reset each day but the data for each day is still available in the database. Being able to look through the previous dates will be implemented in the future. The user can send the customer a text letting them know their table is ready.

The app uses Material UI to manage the UI/UX with Material-Table to  build the tables that are used.

The app currently uses an API that was made to manage sending text messages. That server will likely be taken down in the future to switch to cloud functions with firebase.