The app has been setup to deploy to cloud hosting through Google and the functions have been updated to utilize cloud functions instead of a separate backend.
The LoadingSpinner section of App.js has been temporarily commented out and rewritten to launch a beta of this.

Known issue:
Sending a message throws an internal error. Preliminary investigation shows this may be related to an unhandled promise rejection. The app still currently works but it ends the sendWaitMessage saga early and doesn't send a message success like it should.